+++ chain/beacon/store_test.go
@@ -11,11 +11,12 @@ import (
 	"github.com/drand/drand/chain/boltdb"
 	"github.com/drand/drand/crypto"
 	"github.com/drand/drand/test"
+	context2 "github.com/drand/drand/test/context"
 )
 
 func TestSchemeStore(t *testing.T) {
 	dir := t.TempDir()
-	ctx, sch, _ := test.PrevSignatureMatersOnContext(t, context.Background())
+	ctx, sch, _ := context2.PrevSignatureMatersOnContext(t, context.Background())
 
 	l := test.Logger(t)
 	bstore, err := boltdb.NewBoltStore(ctx, l, dir, nil)
