+++ chain/boltdb/trimmed_test.go
@@ -10,12 +10,13 @@ import (
 	"github.com/drand/drand/chain"
 	chainerrors "github.com/drand/drand/chain/errors"
 	"github.com/drand/drand/test"
+	context2 "github.com/drand/drand/test/context"
 )
 
 func TestTrimmedStoreBoltOrder(t *testing.T) {
 	tmp := t.TempDir()
 
-	ctx, _, prevMatters := test.PrevSignatureMatersOnContext(t, context.Background())
+	ctx, _, prevMatters := context2.PrevSignatureMatersOnContext(t, context.Background())
 
 	if prevMatters {
 		// This test stores b2 then b1. However, when the beacon order matters, the correct
@@ -83,7 +84,7 @@ func TestTrimmedStoreBoltOrder(t *testing.T) {
 func TestTrimmedStoreBolt(t *testing.T) {
 	tmp := t.TempDir()
 
-	ctx, _, prevMatters := test.PrevSignatureMatersOnContext(t, context.Background())
+	ctx, _, prevMatters := context2.PrevSignatureMatersOnContext(t, context.Background())
 
 	l := test.Logger(t)
 
@@ -201,7 +202,7 @@ func TestTrimmedStoreBolt(t *testing.T) {
 func TestTrimmedStore_Cursor(t *testing.T) {
 	tmp := t.TempDir()
 
-	ctx, _, prevMatters := test.PrevSignatureMatersOnContext(t, context.Background())
+	ctx, _, prevMatters := context2.PrevSignatureMatersOnContext(t, context.Background())
 
 	l := test.Logger(t)
 	dbStore, err := newTrimmedStore(ctx, l, tmp, nil)
