+++ core/drand_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/drand/drand/protobuf/common"
 	"github.com/drand/drand/protobuf/drand"
 	"github.com/drand/drand/test"
+	context2 "github.com/drand/drand/test/context"
 )
 
 func setFDLimit(t *testing.T) {
@@ -954,7 +955,7 @@ func TestDrandCheckChain(t *testing.T) {
 		t.Skip(`This test does not work with in-memory database. See the "// Skip why: " comment for details.`)
 	}
 
-	ctx, _, prevMatters := test.PrevSignatureMatersOnContext(t, context.Background())
+	ctx, _, prevMatters := context2.PrevSignatureMatersOnContext(t, context.Background())
 
 	n, p := 4, 1*time.Second
 	beaconID := test.GetBeaconIDFromEnv()
