+++ test/test.go
@@ -3,18 +3,13 @@
 package test
 
 import (
-	"context"
 	"encoding/hex"
 	n "net"
 	"os"
 	"strconv"
 	"sync"
-	"testing"
 	"time"
 
-	"github.com/stretchr/testify/require"
-
-	"github.com/drand/drand/chain"
 	commonutils "github.com/drand/drand/common"
 	"github.com/drand/drand/crypto"
 	"github.com/drand/drand/key"
@@ -188,22 +183,3 @@ func GetBeaconIDFromEnv() string {
 	return commonutils.GetCanonicalBeaconID(beaconID)
 }
 
-// PrevSignatureMatersOnContext checks if the previous signature matters or not for future operations
-// based on the used schema.
-// If it matters, then it's also set on the passed context.
-func PrevSignatureMatersOnContext(t *testing.T, ctx context.Context) (context.Context, *crypto.Scheme, bool) {
-	sch, err := crypto.GetSchemeFromEnv()
-	require.NoError(t, err)
-
-	prevMatters := true
-	if sch.Name == crypto.UnchainedSchemeID ||
-		sch.Name == crypto.ShortSigSchemeID {
-		prevMatters = false
-	}
-
-	if prevMatters {
-		ctx = chain.SetPreviousRequiredOnContext(ctx)
-	}
-
-	return ctx, sch, prevMatters
-}
